Formatting SSTV seems to be working...
I am comparing my output to the output of the pisstv program. It's
identical for a pure red PNG file, but the pisstv seems to output one
too many rows? My program is identical up to that point though. Still
need to test on a Rasperry Pi and a with a radio scanner.